Navigation bar implement
Navigation bar is now implemented

It will display on all expect pages

for example

Dashboard

Moive.Index

Movie.Show

For guest, it will show 'login' and 'register' in the top right corner of the screen. Once they are logged in, these elements will be replaced by the users full name, with the option of a drop menu, currently with a single item of 'log out'

Further changes are still needing to be made to the navigation bar; namely, when the user logins in from movie.index or movie.show, it will redirect them to dashboard page. Need to change it so it leaves the user on the current page that they were in