Add and edit pages updated
Custom layout has been created for Add and Edit page.

Functions are all identical as before

Minor responsive implemented: Reduces columns of contents from 2 to 1 so all input text box appear in a single file

Added in Celebrity and Genre TextArea for future Many to Many function implementation. Unsure if this is the best approach to take (how to ensure that movie or genre provide by Admin exists in database? Does a custom requirement need to be defined?)